Time: 0 ms (100.00%) | Memory: 113.7 MB (61.37%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.cpp
@@ -1,12 +1,12 @@
 class Solution {
 public:
     double findMaxAverage(vector<int>& nums, int k) {
-        long long windowSum = 0;
+        int windowSum = 0;
         //made sliding window of k elements
         for(int i = 0; i<k; i++){
             windowSum+=nums[i];
         }
-        long long maxSum = windowSum;
+        int maxSum = windowSum;
 
         //add next eleemnt to k and remove 0th element, find maximumsum and return it.
         for(int right = k; right<nums.size(); right++){
